src/requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
cffi==1.14.2
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==3.1
cssselect2==0.3.0
daphne==2.5.0
defusedxml==0.6.0
Django==3.1
django-extensions==2.2.5
flower==0.9.3
hiredis==1.1.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
kombu==4.6.11
msgpack==0.6.2
Pillow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.0.0
Pyphen==0.9.5
PyQRCode==1.2.1
pytz==2020.1
pyzbar==0.1.8
qrcode==6.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tinycss2==1.0.2
tornado==5.1.1
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.16.0
zipp==3.1.0
zope.interface==5.1.0
psycopg2-binary==2.8.6
whitenoise
gunicorn
dj-database-url